+++ src/web.ts
@@ -1,6 +1,6 @@
 import { WebPlugin } from '@capacitor/core';
 
-import type { HealthConnectPlugin, PermissionsResponse, RecordType } from './definitions';
+import type { HealthConnectPlugin, PermissionsResponse, RecordType, AggregateRecordType, AggregateResponse, AggregateGroupBy } from './definitions';
 
 export class HealthConnectWeb extends WebPlugin implements HealthConnectPlugin {
   async checkAvailability(): Promise<{ availability: 'Available' | 'NotSupported' | 'NotInstalled' }> {
@@ -27,5 +27,11 @@ export class HealthConnectWeb extends WebPlugin implements HealthConnectPlugin {
 
   async readRecords(options: { start: string; end: string; type: RecordType }): Promise<any> {
     console.warn('HealthConnect is not available on the web', options);
+  }
+
+  async aggregateRecords(options: { start: string; end: string; type: AggregateRecordType; groupBy?: AggregateGroupBy }): Promise<AggregateResponse> {
+    console.warn('HealthConnect is not available on the web', options);
+    
+    return { aggregates: [] };
   }
 }